Move `moveCamera()` to SceneContext.jsx. && Dev.
