+++ src/components/Selector.jsx
@@ -153,6 +153,7 @@ export default function Selector({templateInfo, animationManager, blinkManager, 
   },[selectedOptions])
   // user selects an option
   const selectTraitOption = (option) => {
+    debugger
     if (option == null){
       option = {
         item:null,
@@ -553,6 +554,11 @@ export default function Selector({templateInfo, animationManager, blinkManager, 
   }
 
   function ClearTraitButton() {
+    // if (removeOption) {
+    //   console.log('-ClearTraitButton 1')
+    // } else {
+    //   console.log('-ClearTraitButton 2')
+    // }
     // clear the current trait
     return removeOption ? (
       <div
