+++ src/context/SceneContext.jsx
@@ -4,6 +4,7 @@ import * as THREE from "three"
 import {
   getRandomizedTemplateOptions
 } from "../library/option-utils"
+import gsap from "gsap"
 
 export const SceneContext = createContext()
 
@@ -98,6 +99,46 @@ export const SceneProvider = (props) => {
     setAvatar({})
   }
 
+  const moveCamera = (value) => {
+    if (!controls) return
+    gsap.to(controls.target, {
+      y: value.height,
+      x: 0,
+      z: 0,
+      duration: 1,
+    })
+
+    gsap
+      .fromTo(
+        controls,
+        {
+          maxDistance: controls.getDistance(),
+          minDistance: controls.getDistance(),
+          minPolarAngle: controls.getPolarAngle(),
+          maxPolarAngle: controls.getPolarAngle(),
+          minAzimuthAngle: controls.getAzimuthalAngle(),
+          maxAzimuthAngle: controls.getAzimuthalAngle(),
+        },
+        {
+          maxDistance: value.distance,
+          minDistance: value.distance,
+          minPolarAngle: Math.PI / 2 - 0.11,
+          maxPolarAngle: Math.PI / 2 - 0.11,
+          minAzimuthAngle: -0.78,
+          maxAzimuthAngle: -0.78,
+          duration: 1,
+        },
+      )
+      .then(() => {
+        controls.minPolarAngle = 0
+        controls.maxPolarAngle = 3.1415
+        controls.minDistance = 0.5
+        controls.maxDistance = 5
+        controls.minAzimuthAngle = Infinity
+        controls.maxAzimuthAngle = Infinity
+      })
+  }
+
   return (
     <SceneContext.Provider
       value={{
@@ -141,6 +182,7 @@ export const SceneProvider = (props) => {
         avatar,
         setAvatar,
         resetAvatar,
+        moveCamera,
         traitsNecks,
         setTraitsNecks,
         traitsSpines,
